Adding an extensible Node server which libraries can define endpoints for specific functionality.
The server, django_node.server.NodeServer, sits alongside the python process and accepts requests via HTTP. The goal here is to provide a means for utility tasks to be completed without the delay on booting node and loading the required libraries.

The benefit of this approach was proven when django-react switched to a persistent render server and experienced a massive performance boost. The intention here is to provide the same boost for other tasks, such as compilation, compression, etc.