django_node.npm.install now pipes the stdout directly to the terminal.
This improves the clarity of run-time dependency checks.

Excising a bunch of internal docs which are well out of sync with the README's. In the long-term it would be good to rely on docstrings, but it is better to have one source of truth.

npm.install will no longer raise a NpmInstallError.

Fixes #7